fix(comp:table): refactor table column width measure logic 1. measured width should be used in table header fix holder 2. table content colgroup width is set using width config in columns
